requirements.txt
Django==2.1
django-rest-framework==0.1.0
django-webpack-loader==0.6.0
djangorestframework==3.8.2
pytz==2018.5
